wip(refactor): create score trait for game state
